fix(code-review): renamed button-new to cb-button
