+++ frontend/src/main.ts
@@ -12,7 +12,7 @@ import store from './store/store';
 import createRouter from './router';
 
 import App from './App.vue';
-import ButtonNew from '@/components/ButtonNew.vue';
+import CBButton from '@/components/CBButton.vue';
 
 import {
   raid as featureFlagRaid,
@@ -35,7 +35,7 @@ Vue.use(VueRouter);
 Vue.use(VTooltip);
 Vue.use(BootstrapVue);
 Vue.use(BootstrapVueDialog);
-Vue.component('ButtonNew', ButtonNew);
+Vue.component('cb-button', CBButton);
 
 Vue.use(BootstrapVueIcons);
